Big pool update.

+++ code/game/machinery/poolcontroller.dm
@@ -10,14 +10,14 @@
 	var/amount_per_transfer_from_this = 5
 	var/list/linkedturfs = list() //List contains all of the linked pool turfs to this controller, assignment happens on New()
 	var/temperature = "normal" //The temperature of the pool, starts off on normal, which has no effects.
-	var/temperaturecolor = "" //used for nanoUI fancyness
 	var/srange = 5 //The range of the search for pool turfs, change this for bigger or smaller pools.
 	var/linkedmist = list() //Used to keep track of created mist
 	var/misted = 0 //Used to check for mist.
 	var/chem_volume = 10
 	var/datum/wires/poolcontroller/wires = null
-	var/open_panel = 0
-	var/blinkwire = 0 //possibly a drain function in the future. Currently just a blinking light.
+	var/drainable = 0
+	var/drained = 0
+	var/timer = 0 //we need a cooldown on that shit.
 	var/seconds_electrified = 0	//Shock morons, like an airlock.
 
 /obj/machinery/poolcontroller/New() //This proc automatically happens on world start
@@ -91,49 +91,52 @@
 /obj/machinery/poolcontroller/process()
 	updatePool() //Call the mob affecting proc
 	poison()
+	if(timer > 0)
+		timer--
 
 /obj/machinery/poolcontroller/proc/updatePool()
-	for(var/turf/simulated/pool/water/W in linkedturfs) //Check for pool-turfs linked to the controller.
-		for(var/mob/M in W) //Check for mobs in the linked pool-turfs.
-			if(isobserver(M))
-				continue
+	if(!drained)
+		for(var/turf/simulated/pool/water/W in linkedturfs) //Check for pool-turfs linked to the controller.
+			for(var/mob/M in W) //Check for mobs in the linked pool-turfs.
+				if(isobserver(M))
+					continue
 
-			//End sanity checks, go on
-			switch(temperature) //Apply different effects based on what the temperature is set to.
-				if("scalding") //Burn the mob.
-					M.bodytemperature = min(500, M.bodytemperature + 35) //heat mob at 35k(elvin) per cycle
-					// if(M.bodytemperature >= 400 && !M.stat)
-						// M << "<span class='danger'>You're boiling alive!</span>"
+				//End sanity checks, go on
+				switch(temperature) //Apply different effects based on what the temperature is set to.
+					if("Scalding") //Burn the mob.
+						M.bodytemperature = min(500, M.bodytemperature + 35) //heat mob at 35k(elvin) per cycle
+						// if(M.bodytemperature >= 400 && !M.stat)
+							// M << "<span class='danger'>You're boiling alive!</span>"
 
-				if("frigid") //Freeze the mob.
-					M.bodytemperature = max(80, M.bodytemperature - 35) //cool mob at -35k per cycle, less would not affect the mob enough.
-					// if(M.bodytemperature <= 215 && !M.stat)
-						// M << "<span class='danger'>You're being frozen solid!</span>"
+					if("Frigid") //Freeze the mob.
+						M.bodytemperature = max(80, M.bodytemperature - 35) //cool mob at -35k per cycle, less would not affect the mob enough.
+						// if(M.bodytemperature <= 215 && !M.stat)
+							// M << "<span class='danger'>You're being frozen solid!</span>"
 
-				if("normal") //Normal temp does nothing, because it's just room temperature water.
+					if("Normal") //Normal temp does nothing, because it's just room temperature water.
 
-				if("warm") //Gently warm the mob.
-					M.bodytemperature = min(360, M.bodytemperature + 20) //Heats up mobs till the termometer shows up
+					if("Warm") //Gently warm the mob.
+						M.bodytemperature = min(360, M.bodytemperature + 20) //Heats up mobs till the termometer shows up
 
-				if("cool") //Gently cool the mob.
-					M.bodytemperature = max(250, M.bodytemperature - 20) //Cools mobs till the termometer shows up
-			var/mob/living/carbon/human/drownee = M
-			if(drownee.stat == DEAD)
-				continue
-			if(drownee && drownee.lying && !drownee.internal)
-				if(drownee.stat != CONSCIOUS)
-					drownee.adjustOxyLoss(9)
-					drownee << "<span class='danger'>You're quickly drowning!</span>"
-				else
-					if(!drownee.internal)
-						drownee.adjustOxyLoss(5)
-						if(prob(35))
-							drownee << "<span class='danger'>You're lacking air!</span>"
+					if("Cool") //Gently cool the mob.
+						M.bodytemperature = max(250, M.bodytemperature - 20) //Cools mobs till the termometer shows up
+				var/mob/living/carbon/human/drownee = M
+				if(drownee.stat == DEAD)
+					continue
+				if(drownee && drownee.lying && !drownee.internal)
+					if(drownee.stat != CONSCIOUS)
+						drownee.adjustOxyLoss(9)
+						drownee << "<span class='danger'>You're quickly drowning!</span>"
+					else
+						if(!drownee.internal)
+							drownee.adjustOxyLoss(5)
+							if(prob(35))
+								drownee << "<span class='danger'>You're lacking air!</span>"
 
-		for(var/obj/effect/decal/cleanable/decal in W)
-			animate(decal, alpha = 10, time = 20)
-			spawn(25)
-				qdel(decal)
+			for(var/obj/effect/decal/cleanable/decal in W)
+				animate(decal, alpha = 10, time = 20)
+				spawn(25)
+					qdel(decal)
 
 
 /obj/machinery/poolcontroller/proc/miston() //Spawn /obj/effect/mist (from the shower) on all linked pool tiles
@@ -155,14 +158,41 @@
 	var/dat = ""
 	if(panel_open)
 		dat += wires()
-	if(emagged)
-		dat += "<a href='?src=\ref[src];Scalding=1'>Scalding</a><br>"
-	dat += "<a href='?src=\ref[src];Warm=1'>Warm</a><br>"
-	dat += "<a href='?src=\ref[src];Normal=1'>Normal</a><br>"
-	dat += "<a href='?src=\ref[src];Cool=1'>Cool</a><br>"
-	if(emagged)
-		dat += "<a href='?src=\ref[src];Frigid=1'>Frigid</a><br>"
-	var/datum/browser/popup = new(user, "Pool Controller", name, 350, 480)
+
+
+	dat += text({"
+				<BR><TT><B>Pool Controller Panel</B></TT><BR><BR>
+				<B>Current temperature :</B> [temperature]<BR>
+				</B>Drain is [drainable ? "active" : "unactive"]</B> <BR>"})
+
+	dat += text({"<B>Pool status :</B>      "})
+	if(timer < 45)
+		switch(drained)
+			if(0)
+				dat += "<span class='good'>Full</span><BR>"
+			if(1)
+				dat += "<span class='bad'>Drained</span><BR>"
+	if(timer >= 45)
+		dat += "<span class='bad'>Warning. Do not approach drain!<BR></span>"
+
+	dat += "<span class='notice'>[timer] seconds left until pool can operate again.<BR></span>"
+
+	if(timer == 0 && drained == 0)
+		if(emagged)
+			dat += "<a href='?src=\ref[src];Scalding=1'>Scalding</a><br>"
+		dat += "<a href='?src=\ref[src];Warm=1'>Warm</a><br>"
+		dat += "<a href='?src=\ref[src];Normal=1'>Normal</a><br>"
+		dat += "<a href='?src=\ref[src];Cool=1'>Cool</a><br>"
+		if(emagged)
+			dat += "<a href='?src=\ref[src];Frigid=1'>Frigid</a><br>"
+	if(drainable && !drained && timer == 0)
+		dat += "<a href='?src=\ref[src];Activate Drain=1'>Drain Pool</a><br>"
+	if(drainable && drained && timer == 0)
+		dat += "<a href='?src=\ref[src];Activate Drain=1'>Fill Pool</a><br>"
+
+
+
+	var/datum/browser/popup = new(user, "Pool Controller", name, 350, 520)
 	popup.set_content(dat)
 	popup.open()
 
@@ -173,29 +203,56 @@
 		return
 	if(href_list["Scalding"])
 		if(emagged)
-			src.temperature = "scalding"
+			timer = 10
+			src.temperature = "Scalding"
 			src.icon_state = "poolcscald"
-			miston()
+			if(!misted)
+				miston()
 		if(!emagged)
 			message_admins("[key_name_admin(usr)] is trying to use href exploits with the poolcontroller")
 	if(href_list["Warm"])
-		src.temperature = "warm"
+		timer = 10
+		src.temperature = "Warm"
 		src.icon_state = "poolcwarm"
 		mistoff()
 	if(href_list["Normal"])
-		src.temperature = "normal"
+		timer = 10
+		src.temperature = "Normal"
 		src.icon_state = "poolcnorm"
 		mistoff()
 	if(href_list["Cool"])
-		src.temperature = "cool"
+		timer = 10
+		src.temperature = "Cool"
 		src.icon_state = "poolccool"
 		mistoff()
 	if(href_list["Frigid"])
 		if(emagged)
-			src.temperature = "frigid"
+			timer = 10
+			src.temperature = "Frigid"
 			src.icon_state = "poolcfrig"
 			mistoff()
 		if(!emagged)
+			message_admins("[key_name_admin(usr)] is trying to use href exploits with the poolcontroller")
+	if(href_list["Activate Drain"])
+		if(drainable) //Drain is activated
+			for(var/obj/machinery/drain/drain in range(6, src))
+				if(drain.active == 1)
+					return
+				if(timer > 0)
+					return
+				else
+					mistoff()
+					drain.active = 1
+					drain.timer = 15
+					timer = 60
+					if(drain.status == 0)
+						new /obj/effect/whirlpool(drain.loc)
+						src.temperature = "None"
+						src.icon_state = "poolcnorm"
+					if(drain.status == 1)
+						new /obj/effect/effect/waterspout(drain.loc)
+						src.temperature = "Normal"
+		if(!drainable)
 			message_admins("[key_name_admin(usr)] is trying to use href exploits with the poolcontroller")
 	update_icon()
 	updateUsrDialog()
